Refactor AccessControlLayer and VotingEscrowMoca contracts: - Change MONITOR_ROLE visibility to private and add GLOBAL_ADMIN role. - Introduce functions for managing GLOBAL_ADMIN role in AccessControlLayer. - Update VotingEscrowMoca to use IAddressBook for address management and remove AccessControl inheritance. - Implement new modifiers for role checks and enhance epoch handling in various functions. - Optimize gas usage and improve code readability by streamlining mapping lookups and function signatures.

+++ src/libraries/EpochMath.sol
@@ -35,9 +35,9 @@ library EpochMath {
     }
     
 
-    ///@dev returns current epoch start time
-    function getCurrentEpochStart() internal view returns (uint256) {
-        return getEpochStartForTimestamp(block.timestamp);
+    ///@dev returns current epoch start time | uint128: Checkpoint{veBla, uint128 lastUpdatedAt}
+    function getCurrentEpochStart() internal view returns (uint128) {
+        return uint128(getEpochStartForTimestamp(block.timestamp));
     }
 
     ///@dev returns current epoch end time
@@ -51,8 +51,7 @@ library EpochMath {
         return (epoch + 1) * EPOCH_DURATION;
     }
 
-// ------------ needed?
-
+    // used in _createLockFor()
     function isValidEpochTime(uint256 timestamp) internal pure returns (bool) {
         return timestamp % EPOCH_DURATION == 0;
     }
